+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoPlayer.kt
@@ -31,7 +31,7 @@ import java.lang.ref.WeakReference
 
 // https://developer.android.com/guide/topics/media/media3/getting-started/migration-guide#improvements_in_media3
 @UnstableApi
-class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?) : AutoCloseable, SharedObject(appContext) {
+class VideoPlayer(val context: Context, appContext: AppContext, source: VideoSource?) : AutoCloseable, SharedObject(appContext) {
   // This improves the performance of playing DRM-protected content
   private var renderersFactory = DefaultRenderersFactory(context)
     .forceEnableMediaCodecAsynchronousQueueing()
@@ -236,8 +236,8 @@ class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?
 
   fun prepare() {
     uncommittedSource?.let { videoSource ->
-      val mediaItem = videoSource.toMediaItem()
-      player.setMediaItem(mediaItem)
+      val mediaSource = videoSource.toMediaSource(context)
+      player.setMediaSource(mediaSource)
       player.prepare()
       lastLoadedSource = videoSource
       uncommittedSource = null