+++ packages/expo-video/src/VideoPlayer.types.ts
@@ -178,6 +178,13 @@ export type VideoSource =
        * When undefined the player will display information contained in the video metadata.
        */
       metadata?: VideoMetadata;
+      /**
+       * Specifies headers sent with the video request.
+       * > For DRM license headers use the `headers` field of [`DRMOptions`](#drmoptions).
+       * @platform android
+       * @platform ios
+       */
+      headers?: Record<string, string>;
     }
   | null;